+++ src/main/java/com/structurizr/dsl/StructurizrDslTokens.java
@@ -23,7 +23,7 @@ class StructurizrDslTokens {
     static final String WORKSPACE_TOKEN = "workspace";
     static final String EXTENDS_TOKEN = "extends";
     static final String REF_TOKEN = "ref";
-    static final String IMPLIED_RELATIONSHIPS_TOKEN = "impliedRelationships";
+    static final String IMPLIED_RELATIONSHIPS_TOKEN = "!impliedRelationships";
     static final String MODEL_TOKEN = "model";
     static final String VIEWS_TOKEN = "views";
     static final String ENTERPRISE_TOKEN = "enterprise";