+++ src/test/java/com/structurizr/dsl/ImpliedRelationshipsParserTests.java
@@ -10,22 +10,44 @@ class ImpliedRelationshipsParserTests extends AbstractTests {
     private ImpliedRelationshipsParser parser = new ImpliedRelationshipsParser(false);
 
     @Test
-    void test_parse_ThrowsAnException_WhenThereAreTooManyTokens() {
+    void test_parse_ThrowsAnException_WhenThereAreTooManyTokensAndNotRestricted() {
         try {
             parser.parse(context(), tokens("impliedRelationships", "boolean", "extra"));
             fail();
         } catch (Exception e) {
-            assertEquals("Too many tokens, expected: impliedRelationships <true|false|fqn>", e.getMessage());
+            assertEquals("Too many tokens, expected: !impliedRelationships <true|false|fqn>", e.getMessage());
         }
     }
 
     @Test
-    void test_parse_ThrowsAnException_WhenNoFlagIsSpecified() {
+    void test_parse_ThrowsAnException_WhenThereAreTooManyTokensAndRestricted() {
+        try {
+            parser = new ImpliedRelationshipsParser(true);
+            parser.parse(context(), tokens("impliedRelationships", "boolean", "extra"));
+            fail();
+        } catch (Exception e) {
+            assertEquals("Too many tokens, expected: !impliedRelationships <true|false>", e.getMessage());
+        }
+    }
+
+    @Test
+    void test_parse_ThrowsAnException_WhenNoFlagIsSpecifiedAndNotRestricted() {
         try {
             parser.parse(context(), tokens("impliedRelationships"));
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: impliedRelationships <true|false|fqn>", e.getMessage());
+            assertEquals("Expected: !impliedRelationships <true|false|fqn>", e.getMessage());
+        }
+    }
+
+    @Test
+    void test_parse_ThrowsAnException_WhenNoFlagIsSpecifiedAndRestricted() {
+        try {
+            parser = new ImpliedRelationshipsParser(true);
+            parser.parse(context(), tokens("impliedRelationships"));
+            fail();
+        } catch (Exception e) {
+            assertEquals("Expected: !impliedRelationships <true|false>", e.getMessage());
         }
     }
 
@@ -52,10 +74,13 @@ class ImpliedRelationshipsParserTests extends AbstractTests {
 
     @Test
     void test_parse_SetsTheStrategy_WhenAFullyQualifiedClassNameIsSpecifiedAndRestricted() {
-        parser = new ImpliedRelationshipsParser(true);
-        parser.parse(context(), tokens("impliedRelationships", "com.structurizr.model.CreateImpliedRelationshipsUnlessSameRelationshipExistsStrategy"));
-
-        assertEquals("com.structurizr.model.CreateImpliedRelationshipsUnlessAnyRelationshipExistsStrategy", workspace.getModel().getImpliedRelationshipsStrategy().getClass().getCanonicalName());
+        try {
+            parser = new ImpliedRelationshipsParser(true);
+            parser.parse(context(), tokens("impliedRelationships", "com.structurizr.model.CreateImpliedRelationshipsUnlessSameRelationshipExistsStrategy"));
+            fail();
+        } catch (Exception e) {
+            assertEquals("Custom implied relationship strategies are not available when running in restricted mode, expected: !impliedRelationships <true|false>", e.getMessage());
+        }
     }
 
 }